fix DecodeConfig in dynamic mode too
Like #9, but for dynamic mode.

And more tests. (including exporting a bool to let tests in callers
like Perkeep exercise both the wasm + dynamic code paths)

This also makes adds the common macOS homebrew libheif path as a
location that's tried in dynamic mode, and installs it for tests in
CI.

This also disables dynamic mode on Windows, as it doesn't work yet.

Updates #9
Updates #11

Signed-off-by: Brad Fitzpatrick <brad@danga.com>